examples / adf_show_metadata_volume: int types clean-up.

+++ examples/adf_show_metadata_volume.c
@@ -28,9 +28,9 @@ void show_volume_metadata ( struct AdfVolume * const vol )
     //numCyls = highCyl - lowCyl + 1
     //highKey = numCyls * numSurfaces * numBlocksPerTrack - 1
     //rootKey = INT (numReserved + highKey) / 2
-    int root_block_sector = //880,
+    uint32_t root_block_sector = //880,
         //bblock.rootBlock,  // in many images is 0 ???
-        ( ( vol->lastBlock - vol->firstBlock ) / 2 + 1 ); // this seems to work
+        (uint32_t) ( ( vol->lastBlock - vol->firstBlock ) / 2 + 1 ); // this seems to work
     printf ("\nRoot block sector:\t%u\n", root_block_sector );
 
     if ( adfReadRootBlock ( vol, root_block_sector, &rblock ) != RC_OK ) {
